requirements.txt
Django==2.2.16
dj-database-url==0.5.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.1
drf-writable-nested==0.6.2
psycopg2-binary==2.8.6
model-bakery==1.2.0
